requirements.txt
ultralytics
deep-sort-realtime
facenet-pytorch
torchvision
fastapi
uvicorn
opencv-python
numpy
jinja2
python-multipart
onnxruntime
pillow
scikit-learn
